Atualização do projeto

+++ API-SITE-Mulher/Repository/Implementations/GenericRepository.cs
@@ -30,7 +30,7 @@ namespace API_SITE_Mulher.Repository.Implementations
             return item;
         }
 
-        public void DeleteById(int id)
+        public void DeleteById(long id)
         {
             var result = dataset.SingleOrDefault((t) => t.Id.Equals(id));
             if (result != null)
